assigns/untitled/fall-2020-itws-4250-6250-master/examples/python-venv-lab/requirements.txt
psycopg2-binary
tabulate